refactor: move q8 packed runtime gates
